Initial implementation of EditorConfig Lua bindings
The C API editorconfig_parse() function is made available in the
parse() function of the "editorconfig_core" module. It returns
a table with each key as a name string and the value as a Lua
type. Many values for known property names are tokens defined
in the module namespace (instead of plain EC value strings).